requirements.txt
# Install the PyTorch build that matches your CUDA driver separately, then install:
ultralytics
numpy
pyyaml
pandas
matplotlib
seaborn
pillow
onnx
onnxruntime
